+++ src/Sim/BoostPad/BoostPadGrid/BoostPadGrid.h
@@ -15,8 +15,8 @@ struct BoostPadGrid {
 	constexpr static int
 		CELLS_X = 8,
 		CELLS_Y = 10,
-		CELL_SIZE_X = EXTENT_X / (CELLS_X / 2),
-		CELL_SIZE_Y = EXTENT_Y / (CELLS_Y / 2),
+		CELL_SIZE_X = (int)(EXTENT_X / (CELLS_X / 2)),
+		CELL_SIZE_Y = (int)(EXTENT_Y / (CELLS_Y / 2)),
 		CELL_AMOUNT = CELLS_X * CELLS_Y;
 
 	BoostPad* pads[CELLS_X][CELLS_Y] = {};
